+++ docs/ReleaseNotes.md
@@ -44,6 +44,8 @@ line upon naming the release. Refer to previous for appropriate section names.
 - The `shared` and `uniform` keywords are removed in HLSL 202x, with
   compatibility warnings available for earlier language versions
   [#8482](https://github.com/microsoft/DirectXShaderCompiler/issues/8482).
+- Added support for `auto` return types for normal functions aligning with C++14
+  [#8903](https://github.com/microsoft/DirectXShaderCompiler/issues/8903).
 - HLSL 202x now supports variadic templates and template function parameter
   packs, pack expansions, and `sizeof...()`
   [#8905](https://github.com/microsoft/DirectXShaderCompiler/issues/8905).